DEVELOPER-1072 video description trucation
We're truncating a bit differently in aweplug now. Before we were
truncating at the end of a sentence, as long as it was less than 150
characters. Now we're just truncating at 150. Ideally each video would
have a sentence or two that fit within that 150 range and we would have
been fine, but that is not the case.