Updating aweplug for the issues with angular
You can look at
https://github.com/awestruct/aweplug/commit/8a2df4485405593296c37eaf4e6361ef3c1c8fd9
for all the information in this change.

I rearranged the locations of the aweplug dependency is the Gemfile to
put all the various aweplug locations in the same area of the file. I
also made activesupport a hard dependency so we can manage the version
ourselves instead of ruby-duration.