Fixes DEVELOPER-491 (contributor widget)
There's a lot going on here. Firstly, we now have a javascript widget we
can use anywhere we want the contributor widget to show. Here are the
steps to use it:

* Add in the foundation/foundation.dropdown.js javascript to the page
* Create a placeholder `span[data-sys-contributor="Name <email>"]`
* Create an array of contributors on the page
* Call app.dcp.resolveContributors with the array of contributors

This is now in place for all of the places I know about where we show
contributors / authors: books, developer materials, videos, quickstarts,
and demos.

If changes to the widget need to be made they need to be made in both
paths of the `_partial/contributor.html.slim` file, or possibly in the
app.scss to change the styling.

I added a new `Object.keys` polyfill for older browsers.