Fixes DEVEOLPER-1376 (Fixing the build locally)
The local build was trying to use the JDG cache without any info about
where it is, so you ended up doing multiple hits to the cache which
would error out then you'd have to do the full request.

I'm not sure when this changed, but things should be back to normal.